+++ src/GraphicalInterface.py
@@ -1,4 +1,4 @@
-APP_VERSION = "1.0.15"
+APP_VERSION = "1.0.16"
 """
 GraphicalInterface.py - Main Application Window for ErgoProtect
 ----------------------------------------------------------------